+++ Sora/Sora.cs
@@ -70,6 +70,7 @@ public class Sora : IDisposable
     {
         OPUS,
     }
+
     // SimulcastRid のためのパラメータ
     public enum SimulcastRidType
     {
@@ -85,6 +86,17 @@ public class Sora : IDisposable
         R0,
         R1,
         R2,
+    }
+    public enum DegradationPreference
+    {
+        // degradation preference を無効にします
+        Disabled,
+        // フレームレートを優先して維持し、帯域不足時は主に解像度を下げます
+        MaintainFramerate,
+        // 解像度を優先して維持し、帯域不足時は主にフレームレートを下げます
+        MaintainResolution,
+        // 解像度とフレームレートの両方をバランスよく下げます。
+        Balanced,
     }
     public class DataChannel
     {
@@ -369,6 +381,7 @@ public class Sora : IDisposable
         public string VideoAv1Params = "";
         public string VideoH264Params = "";
         public int VideoBitRate = 0;
+        public DegradationPreference? DegradationPreference;
         // デバイスから録音する代わりに Sora.ProcessAudio() で指定したデータを録音データとして利用するかどうか
         public bool UnityAudioInput = false;
         // 再生データをデバイスで再生する代わりに Sora.OnHandleAudio コールバックで再生データを受け取るようにするかどうか
@@ -594,6 +607,24 @@ public class Sora : IDisposable
         cc.video_av1_params = config.VideoAv1Params;
         cc.video_h264_params = config.VideoH264Params;
         cc.video_bit_rate = config.VideoBitRate;
+        if (config.DegradationPreference.HasValue)
+        {
+            switch (config.DegradationPreference.Value)
+            {
+                case DegradationPreference.Disabled:
+                    cc.SetDegradationPreference(SoraConf.Internal.DegradationPreference.DISABLED);
+                    break;
+                case DegradationPreference.MaintainFramerate:
+                    cc.SetDegradationPreference(SoraConf.Internal.DegradationPreference.MAINTAIN_FRAMERATE);
+                    break;
+                case DegradationPreference.MaintainResolution:
+                    cc.SetDegradationPreference(SoraConf.Internal.DegradationPreference.MAINTAIN_RESOLUTION);
+                    break;
+                case DegradationPreference.Balanced:
+                    cc.SetDegradationPreference(SoraConf.Internal.DegradationPreference.BALANCED);
+                    break;
+            }
+        }
         cc.unity_audio_input = config.UnityAudioInput;
         cc.unity_audio_output = config.UnityAudioOutput;
         cc.audio_recording_device = config.AudioRecordingDevice;